finished ex00 and 01

+++ cpp08/ex01/Span.hpp
@@ -11,7 +11,7 @@
 class Span {
 private:
 	Span();
-	int max_size;
+	unsigned long max_size;
 	std::vector<int> vector;
 	int min();
 	int max();
